Fixed #45
* extracted about fragment
* added open source note
* simplified about screen
* restructured license notes
