+++ src/main/java/com/codingspezis/android/metalonly/player/fragments/AboutFragment.java
@@ -1,0 +1,79 @@
+package com.codingspezis.android.metalonly.player.fragments;
+
+import android.content.*;
+import android.net.*;
+import android.os.*;
+import android.support.v4.app.*;
+import android.text.method.*;
+import android.widget.*;
+
+import com.actionbarsherlock.app.*;
+import com.actionbarsherlock.view.*;
+import com.codingspezis.android.metalonly.player.*;
+
+import org.androidannotations.annotations.*;
+import org.androidannotations.annotations.res.*;
+import org.slf4j.*;
+
+@EFragment(R.layout.fragment_about)
+public class AboutFragment extends SherlockFragment {
+    @StringRes
+    String mailaddress_codingspezis, app_name;
+
+    @ViewById
+    TextView textAppVersion;
+
+    @ViewById(R.id.textAboutApp)
+    @FromHtml(R.string.aboutThisApp)
+    TextView mTextAboutApp;
+
+    @ViewById(R.id.textAboutLibraries)
+    @FromHtml(R.string.libraries)
+    TextView mTextAboutLibraries;
+
+    private final Logger LOGGER = LoggerFactory.getLogger(getClass().getSimpleName());
+
+    @AfterViews
+    void linkify() {
+        final MovementMethod linkMovementMethod = LinkMovementMethod.getInstance();
+        mTextAboutApp.setMovementMethod(linkMovementMethod);
+        mTextAboutLibraries.setMovementMethod(linkMovementMethod);
+        if (BuildConfig.DEBUG) LOGGER.debug("linkify() done");
+    }
+
+    /**
+     * sends system intent ACTION_SEND (send mail)
+     *
+     * @param strTo      receiver of mail
+     * @param strSubject subject of mail
+     * @param strText    text of mail
+     */
+    private void sendEmail(String strTo, String strSubject, String strText) {
+        Intent emailIntent = new Intent(Intent.ACTION_SENDTO, Uri.parse("mailto:"));
+        emailIntent.putExtra(Intent.EXTRA_EMAIL, new String[]{strTo});
+        emailIntent.putExtra(Intent.EXTRA_SUBJECT, strSubject);
+        emailIntent.putExtra(Intent.EXTRA_TEXT, strText);
+        try {
+            startActivity(Intent.createChooser(emailIntent, strTo));
+        } catch (android.content.ActivityNotFoundException ex) {
+            MainActivity.toastMessage(getActivity(), getString(R.string.no_mail_app));
+        }
+    }
+
+    @Click(R.id.buttonFeedback)
+    void buttonFeedbackClicked() {
+        String subject = "[" + app_name + " " + BuildConfig.VERSION_NAME
+                + "] Feedback, Fehler";
+        sendEmail(mailaddress_codingspezis, subject, "");
+    }
+
+
+    /**
+     * updates the app version label to the version string generated by gradle
+     */
+    @AfterViews
+    public void setAppVersionLabel() {
+        textAppVersion.setText(BuildConfig.VERSION_NAME);
+    }
+
+}
