+++ src/main/res/layout/fragment_about.xml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+            xmlns:tools="http://schemas.android.com/tools"
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:padding="@dimen/elementMargin">
+
+    <RelativeLayout
+        android:id="@+id/RelativeLayout1"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content">
+
+        <TextView
+            android:id="@+id/textAppName"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentLeft="true"
+            android:layout_alignParentTop="true"
+            android:paddingRight="8dp"
+            android:text="@string/app_name"
+            android:textAppearance="?android:attr/textAppearanceMedium"/>
+
+        <TextView
+            android:id="@+id/textAppVersion"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignBaseline="@+id/textAppName"
+            android:layout_marginTop="@dimen/elementMargin"
+            android:layout_toRightOf="@+id/textAppName"
+            android:text="@string/empty"
+            android:textAppearance="?android:attr/textAppearanceMedium"/>
+
+        <TextView
+            android:id="@+id/textAboutApp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentLeft="true"
+            android:layout_below="@+id/textAppName"
+            android:layout_marginTop="@dimen/elementMarginDouble"
+            android:autoLink="web"
+            android:text="@string/aboutThisApp"/>
+
+        <Button
+            android:id="@+id/buttonFeedback"
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@+id/textAboutApp"
+            android:layout_marginTop="@dimen/elementMargin"
+            android:text="@string/about_mail"/>
+
+        <ImageView
+            android:id="@+id/divider"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:layout_below="@+id/buttonFeedback"
+            android:layout_marginTop="@dimen/elementMargin"
+            android:src="@drawable/grey_horizontal_line"
+            tools:ignore="ContentDescription"/>
+
+        <TextView
+            android:id="@+id/textAboutLibraries"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@+id/divider"
+            android:layout_marginTop="@dimen/elementMarginDouble"
+            android:textAppearance="?android:attr/textAppearanceMedium"/>
+
+    </RelativeLayout>
+
+</ScrollView>
