+++ src/main/res/values/strings_about_activity.xml
@@ -4,7 +4,9 @@
     <!-- about -->
     <string name="mailaddress_codingspezis">mail@codingspezis.com</string>
     <string name="url_codingspezis">http://www.codingspezis.com/</string>
-    <string name="aboutThisApp">Diese App bringt den Internetstream von http://www.metal-only.de auf Android.\nEntwickelt wurde sie von den codingspezis (http://www.codingspezis.com).</string>
+    <string name="aboutThisApp">Diese App bringt den Internetstream von &lt;a href="http://www.metal-only.de>metal-only&lt;/a> auf Android.&lt;br/>
+&lt;br/>
+Entwickelt wird sie hauptsächlich von den &lt;a href="http://www.codingspezis.com">codingspezis&lt;/a> - der Quellcode der App ist auf &lt;a href="https://github.com/Codingspezis/metal-only">github&lt;/a> zu finden. Wir freuen uns sehr über Mit-Entwickler!</string>
     <string name="about_category_app">App</string>
     <string name="about_name">Name:</string>
     <string name="about_version">Version:</string>
@@ -17,42 +19,15 @@
     <string name="about_link">Link:</string>
     <string name="about_license">Lizenz:</string>
 
-    <!-- open core -->
-    <string name="url_opencore">https://android.googlesource.com/platform/external/opencore.git</string>
-    <string name="about_opencore">OpenCORE aacdec</string>
+    <string name="libraries">* &lt;a href="https://android.googlesource.com/platform/external/opencore.git">OpenCORE aacdec&lt;/a>, &lt;a href="https://android.googlesource.com/platform/external/opencore.git/+/master/NOTICE">Apache 2.0&lt;/a>&lt;br/>
+* &lt;a href="https://code.google.com/p/aacdecoder-android/">AAC Decoder for Android&lt;/a>, &lt;a href="http://www.gnu.org/licenses/lgpl.html">LGPL&lt;/a>&lt;br/>
+* &lt;a href="https://github.com/JakeWharton/ActionBarSherlock">ActionBarSherlock&lt;/a>, &lt;a href="https://github.com/JakeWharton/ActionBarSherlock/blob/master/LICENSE.txt">Apache 2.0&lt;/a>&lt;br/>
+* &lt;a href="https://www.github.com/thest1/LazyList">LazyList&lt;/a>, &lt;a href="https://github.com/thest1/LazyList/blob/master/LICENSE">MIT License&lt;/a>&lt;br/>
+* &lt;a href="https://github.com/excilys/androidannotations">androidannotations&lt;/a>, &lt;a href="https://github.com/excilys/androidannotations/blob/develop/LICENSE.txt">Apache 2.0&lt;/a>&lt;br/>
+* &lt;a href="http://www.springsource.org/spring-android">Spring for Android&lt;/a>, &lt;a href="https://github.com/spring-projects/spring-android/blob/master/LICENSE">Apache 2.0&lt;/a>&lt;br/>
+* &lt;a href="http://wiki.fasterxml.com/JacksonHome">Jackson JSON Processor&lt;/a>, &lt;a href="http://wiki.fasterxml.com/JacksonLicensing">Apache 2.0&lt;/a>&lt;br/></string>
 
-    <!-- aac decoder -->
-    <string name="about_aacdecoder">AAC Decoder for Android</string>
-    <string name="url_aacdecoder">http://www.code.google.com/p/aacdecoder-android/</string>
 
-    <!-- actionbarsherlock -->
-    <string name="about_sherlock">ActionBarSherlock</string>
-    <string name="url_sherlock">https://github.com/JakeWharton/ActionBarSherlock/</string>
-
-    <!-- lazylist -->
-    <string name="about_lazylist">LazyList</string>
-    <string name="about_tutorial">Android Custom ListView</string>
-    <string name="url_lazylist">https://www.github.com/thest1/LazyList/</string>
-
-    <!-- androidhive -->
-    <string name="url_androidhive">http://www.androidhive.info/2012/02/android-custom-listview-with-image-and-text/</string>
-
-    <!-- androidannotations -->
-    <string name="about_androidannotations">androidannotations</string>
-    <string name="url_androidannotations">https://github.com/excilys/androidannotations/</string>
-
-    <!-- spring -->
-    <string name="about_spring">Spring for Android</string>
-    <string name="url_spring">http://www.springsource.org/spring-android</string>
-    
-    <!-- jackson -->
-	<string name="about_jackson">Jackson JSON Processor</string>
-	<string name="url_jackson">http://wiki.fasterxml.com/JacksonHome</string>
-
-    <!-- licenses -->
-    <string name="apl2">Apache 2.0</string>
-    <string name="lgpl">LGPL</string>
-    <string name="mit">MIT</string>
     <string name="no_mail_app">Keine E-Mail Anwendung gefunden.</string>
 
 </resources>
